moved helper to decorator and renamed method
