+++ app/helpers/contact_types_helper.rb
@@ -3,23 +3,4 @@ module ContactTypesHelper
   def set_group_options
     @group_options = ContactTypeGroup.for_organization(current_organization).collect { |group| [group.name, group.id] }
   end
-
-  def time_ago_of_last_contact_made_of(contact_type_name, casa_case)
-    contact = last_contact_made_of(contact_type_name, casa_case)
-
-    return nil if contact.nil?
-
-    "#{time_ago_in_words(contact.created_at)} ago"
-  end
-
-  def last_contact_made_of(contact_type_name, casa_case)
-    return unless casa_case
-
-    casa_case
-      .case_contacts
-      .joins(:contact_types)
-      .where(contact_types: {name: contact_type_name})
-      .order(created_at: :desc)
-      .first
-  end
 end
